--- src/components/icon/index.tsx
@@ -1,13 +1,0 @@
-import { FaBeer } from "react-icons/fa";
-import React from "react";
-import { HiHome } from "react-icons/hi2";
-
-export class Question extends React.Component {
-  render() {
-    return (
-      <h3>
-        <HiHome />
-      </h3>
-    );
-  }
-}
