+++ src/container/card/index.tsx
@@ -4,7 +4,7 @@ import { Description } from "../../components/description";
 
 function Card(props: any) {
   return (
-    <div className="container text-center p-3">
+    <div className="container">
       <Title title={props.title} />
       <Description description={props.description} />
       <Picture picture={props.picture} />
