+++ src/pages/contactus.tsx
@@ -1,0 +1,9 @@
+import { Layout } from "../layout"
+
+export function ContactUs() {
+  return (
+    <Layout>
+    <div></div>
+    </Layout>
+  )
+}
