+++ src/pages/galleries.tsx
@@ -1,0 +1,12 @@
+import { Layout } from "../layout"
+
+export function Galleries() {
+
+  return (
+    <Layout>
+    <div>
+
+    </div>
+    </Layout>
+  );
+}
