+++ src/pages/organization.tsx
@@ -1,12 +1,12 @@
-import { KepengurusanContent } from "../content/kepengurusan.content";
+import { OrganizationContent } from "../content/organization.content";
 import { Layout } from "../layout";
 
-export function Kepengurusan() {
+export function Organization() {
   return (
     <Layout>
       <div className="flex flex-col pt-[120px] mb-[50px]">
         <div>
-          <KepengurusanContent />
+          <OrganizationContent />
         </div>
       </div>
     </Layout>
